@@ -1,11 +1,12 @@
 cmake_minimum_required(VERSION 2.8 FATAL_ERROR)
+project(playback)
+message(" ========= Building quizCluster ========== ")
 
-add_definitions(-std=c++11)
+add_definitions(-std=c++14)
 
 set(CXX_FLAGS "-Wall")
 set(CMAKE_CXX_FLAGS, "${CXX_FLAGS}")
 
-project(playback)
 
 find_package(PCL 1.2 REQUIRED)
 
