@@ -3,6 +3,8 @@
 // for exploring self-driving car sensors
 // hahahah
 
+#include <boost/filesystem.hpp>
+
 #include "sensors/lidar.h"
 #include "render/render.h"
 #include "processPointClouds.h"
@@ -11,168 +13,175 @@
 
 std::vector<Car> initHighway(bool renderScene, pcl::visualization::PCLVisualizer::Ptr& viewer)
 {
+  Car egoCar(Vect3(0, 0, 0), Vect3(4, 2, 2), Color(0, 1, 0), "egoCar");
+  Car car1(Vect3(15, 0, 0), Vect3(4, 2, 2), Color(0, 0, 1), "car1");
+  Car car2(Vect3(8, -4, 0), Vect3(4, 2, 2), Color(0, 0, 1), "car2");
+  Car car3(Vect3(-12, 4, 0), Vect3(4, 2, 2), Color(0, 0, 1), "car3");
 
-    Car egoCar( Vect3(0,0,0), Vect3(4,2,2), Color(0,1,0), "egoCar");
-    Car car1( Vect3(15,0,0), Vect3(4,2,2), Color(0,0,1), "car1");
-    Car car2( Vect3(8,-4,0), Vect3(4,2,2), Color(0,0,1), "car2");	
-    Car car3( Vect3(-12,4,0), Vect3(4,2,2), Color(0,0,1), "car3");
-  
-    std::vector<Car> cars;
-    cars.push_back(egoCar);
-    cars.push_back(car1);
-    cars.push_back(car2);
-    cars.push_back(car3);
+  std::vector<Car> cars;
+  cars.push_back(egoCar);
+  cars.push_back(car1);
+  cars.push_back(car2);
+  cars.push_back(car3);
 
-    if(renderScene)
-    {
-        renderHighway(viewer);
-        egoCar.render(viewer);
-        car1.render(viewer);
-        car2.render(viewer);
-        car3.render(viewer);
-    }
+  if (renderScene)
+  {
+    renderHighway(viewer);
+    egoCar.render(viewer);
+    car1.render(viewer);
+    car2.render(viewer);
+    car3.render(viewer);
+  }
 
-    return cars;
+  return cars;
 }
-
 
 void simpleHighway(pcl::visualization::PCLVisualizer::Ptr& viewer)
 {
-    // ----------------------------------------------------
-    // -----Open 3D viewer and display simple highway -----
-    // ----------------------------------------------------
-    
-    // RENDER OPTIONS
-    bool renderScene = false;
-    std::vector<Car> cars = initHighway(renderScene, viewer);
-    
-    // TODO:: Create lidar sensor 
-    Lidar* lidar = new Lidar(cars,0); // Make it on the Heap
-    pcl::PointCloud<pcl::PointXYZ>::Ptr inputCloud = lidar->scan();
-    //renderPointCloud(viewer, inputCloud, "Test");
-    
-    // TODO:: Create point processor
-    ProcessPointClouds<pcl::PointXYZ> pointProcessor; //-- this is on stack
-    //ProcessPointClouds<pcl::PointXYZ>* pointProcessor = new ProcessPointClouds<pcl::PointXYZ>(); // -- this is on heap
-    std::pair<pcl::PointCloud<pcl::PointXYZ>::Ptr, pcl::PointCloud<pcl::PointXYZ>::Ptr> segmentCloud = pointProcessor.SegmentPlane(inputCloud, 100, 0.2);
-    renderPointCloud(viewer, segmentCloud.first, "obstCloud", Color(1,0,0));
-    renderPointCloud(viewer, segmentCloud.second, "planeCloud", Color(0,1,0));
-  
-    std::vector<pcl::PointCloud<pcl::PointXYZ>::Ptr> cloudClusters = pointProcessor.Clustering(segmentCloud.first, 1.0, 3, 30);
+  // ----------------------------------------------------
+  // -----Open 3D viewer and display simple highway -----
+  // ----------------------------------------------------
 
-    int clusterId = 0;
-    std::vector<Color> colors = {Color(1,0,0), Color(1,1,0), Color(0,0,1)};
+  // RENDER OPTIONS
+  bool renderScene = false;
+  std::vector<Car> cars = initHighway(renderScene, viewer);
 
-    for(pcl::PointCloud<pcl::PointXYZ>::Ptr cluster : cloudClusters)
-    {
-        std::cout << "cluster size";
-        pointProcessor.numPoints(cluster);
-        renderPointCloud(viewer, cluster, "obstCloud"+std::to_string(clusterId), colors[clusterId]);
+  // TODO:: Create lidar sensor
+  Lidar* lidar = new Lidar(cars, 0);  // Make it on the Heap
+  pcl::PointCloud<pcl::PointXYZ>::Ptr inputCloud = lidar->scan();
+  // renderPointCloud(viewer, inputCloud, "Test");
 
-        Box box = pointProcessor.BoundingBox(cluster);
-        renderBox(viewer, box, clusterId);
+  // TODO:: Create point processor
+  ProcessPointClouds<pcl::PointXYZ> pointProcessor;  //-- this is on stack
+  // ProcessPointClouds<pcl::PointXYZ>* pointProcessor = new ProcessPointClouds<pcl::PointXYZ>(); // -- this is on heap
+  std::pair<pcl::PointCloud<pcl::PointXYZ>::Ptr, pcl::PointCloud<pcl::PointXYZ>::Ptr> segmentCloud =
+      pointProcessor.SegmentPlane(inputCloud, 100, 0.2);
+  renderPointCloud(viewer, segmentCloud.first, "obstCloud", Color(1, 0, 0));
+  renderPointCloud(viewer, segmentCloud.second, "planeCloud", Color(0, 1, 0));
 
-        ++clusterId;
-    }
+  std::vector<pcl::PointCloud<pcl::PointXYZ>::Ptr> cloudClusters =
+      pointProcessor.Clustering(segmentCloud.first, 1.0, 3, 30);
 
+  int clusterId = 0;
+  std::vector<Color> colors = { Color(1, 0, 0), Color(1, 1, 0), Color(0, 0, 1) };
+
+  for (pcl::PointCloud<pcl::PointXYZ>::Ptr cluster : cloudClusters)
+  {
+    std::cout << "cluster size";
+    pointProcessor.numPoints(cluster);
+    renderPointCloud(viewer, cluster, "obstCloud" + std::to_string(clusterId), colors[clusterId]);
+
+    Box box = pointProcessor.BoundingBox(cluster);
+    renderBox(viewer, box, clusterId);
+
+    ++clusterId;
+  }
 }
 
-
-void cityBlock(pcl::visualization::PCLVisualizer::Ptr& viewer, ProcessPointClouds<pcl::PointXYZI>* pointProcessorI, pcl::PointCloud<pcl::PointXYZI>::Ptr inputCloud)
+void cityBlock(pcl::visualization::PCLVisualizer::Ptr& viewer, ProcessPointClouds<pcl::PointXYZI>* pointProcessorI,
+               pcl::PointCloud<pcl::PointXYZI>::Ptr inputCloud)
 {
-    // ----------------------------------------------------
-    // -----Open 3D viewer and display City Block ---------
-    // ----------------------------------------------------
+  // ----------------------------------------------------
+  // -----Open 3D viewer and display City Block ---------
+  // ----------------------------------------------------
 
-    //ProcessPointClouds<pcl::PointXYZI>* pointProcessorI = new ProcessPointClouds<pcl::PointXYZI>();
-    //pcl::PointCloud<pcl::PointXYZI>::Ptr inputCloud = pointProcessorI->loadPcd("../src/sensors/data/pcd/data_1/0000000000.pcd");
-    //renderPointCloud(viewer, inputCloud, "inputCloud");
-    
-    pcl::PointCloud<pcl::PointXYZI>::Ptr filterCloud = pointProcessorI->FilterCloud(inputCloud, 0.3, Eigen::Vector4f (-10, -5, -2, 1), Eigen::Vector4f (20, 7, 2, 1));
-    //renderPointCloud(viewer, filterCloud, "filterCloud");
+  // ProcessPointClouds<pcl::PointXYZI>* pointProcessorI = new ProcessPointClouds<pcl::PointXYZI>();
+  // pcl::PointCloud<pcl::PointXYZI>::Ptr inputCloud =
+  // pointProcessorI->loadPcd("../src/sensors/data/pcd/data_1/0000000000.pcd"); renderPointCloud(viewer, inputCloud,
+  // "inputCloud");
 
-    std::pair<pcl::PointCloud<pcl::PointXYZI>::Ptr, pcl::PointCloud<pcl::PointXYZI>::Ptr> segmentCloud = pointProcessorI->RansacPlane(filterCloud, 20, 0.2);
-    //renderPointCloud(viewer, segmentCloud.first, "obstCloud", Color(1,0,0));
-    renderPointCloud(viewer, segmentCloud.second, "planeCloud", Color(0,0,1));
-    
-    std::vector<pcl::PointCloud<pcl::PointXYZI>::Ptr> cloudClusters = pointProcessorI->Clustering_Custom(segmentCloud.first, 0.4, 10, 500);
+  pcl::PointCloud<pcl::PointXYZI>::Ptr filterCloud =
+      pointProcessorI->FilterCloud(inputCloud, 0.3, Eigen::Vector4f(-10, -5, -2, 1), Eigen::Vector4f(20, 7, 2, 1));
+  // renderPointCloud(viewer, filterCloud, "filterCloud");
 
-    int clusterId = 0;
-    std::vector<Color> colors = {Color(1,0,0), Color(1,1,0), Color(0,0,1)};
+  std::pair<pcl::PointCloud<pcl::PointXYZI>::Ptr, pcl::PointCloud<pcl::PointXYZI>::Ptr> segmentCloud =
+      pointProcessorI->RansacPlane(filterCloud, 20, 0.2);
+  // renderPointCloud(viewer, segmentCloud.first, "obstCloud", Color(1,0,0));
+  renderPointCloud(viewer, segmentCloud.second, "planeCloud", Color(0, 0, 1));
 
-    for(pcl::PointCloud<pcl::PointXYZI>::Ptr cluster : cloudClusters)
-    {
-        std::cout << "cluster size";
-        pointProcessorI->numPoints(cluster);
-        renderPointCloud(viewer, cluster, "obstCloud"+std::to_string(clusterId), colors[clusterId]);
+  std::vector<pcl::PointCloud<pcl::PointXYZI>::Ptr> cloudClusters =
+      pointProcessorI->Clustering_Custom(segmentCloud.first, 0.4, 10, 500);
 
-        Box box = pointProcessorI->BoundingBox(cluster);
-        renderBox(viewer, box, clusterId);
+  int clusterId = 0;
+  std::vector<Color> colors = { Color(1, 0, 0), Color(1, 1, 0), Color(0, 0, 1) };
 
-        ++clusterId;
-    }
-    
+  for (pcl::PointCloud<pcl::PointXYZI>::Ptr cluster : cloudClusters)
+  {
+    std::cout << "cluster size";
+    pointProcessorI->numPoints(cluster);
+    renderPointCloud(viewer, cluster, "obstCloud" + std::to_string(clusterId), colors[clusterId]);
+
+    Box box = pointProcessorI->BoundingBox(cluster);
+    renderBox(viewer, box, clusterId);
+
+    ++clusterId;
+  }
 }
 
-//setAngle: SWITCH CAMERA ANGLE {XY, TopDown, Side, FPS}
+// setAngle: SWITCH CAMERA ANGLE {XY, TopDown, Side, FPS}
 void initCamera(CameraAngle setAngle, pcl::visualization::PCLVisualizer::Ptr& viewer)
 {
+  viewer->setBackgroundColor(0, 0, 0);
 
-    viewer->setBackgroundColor (0, 0, 0);
-    
-    // set camera position and angle
-    viewer->initCameraParameters();
-    // distance away in meters
-    int distance = 16;
-    
-    switch(setAngle)
-    {
-        case XY : viewer->setCameraPosition(-distance, -distance, distance, 1, 1, 0); break;
-        case TopDown : viewer->setCameraPosition(0, 0, distance, 1, 0, 1); break;
-        case Side : viewer->setCameraPosition(0, -distance, 0, 0, 0, 1); break;
-        case FPS : viewer->setCameraPosition(-10, 0, 0, 0, 0, 1);
-    }
+  // set camera position and angle
+  viewer->initCameraParameters();
+  // distance away in meters
+  int distance = 16;
 
-    if(setAngle!=FPS)
-        viewer->addCoordinateSystem (1.0);
+  switch (setAngle)
+  {
+    case XY:
+      viewer->setCameraPosition(-distance, -distance, distance, 1, 1, 0);
+      break;
+    case TopDown:
+      viewer->setCameraPosition(0, 0, distance, 1, 0, 1);
+      break;
+    case Side:
+      viewer->setCameraPosition(0, -distance, 0, 0, 0, 1);
+      break;
+    case FPS:
+      viewer->setCameraPosition(-10, 0, 0, 0, 0, 1);
+  }
+
+  if (setAngle != FPS)
+    viewer->addCoordinateSystem(1.0);
 }
 
-
-int main (int argc, char** argv)
+int main(int argc, char** argv)
 {
-    std::cout << "starting enviroment" << std::endl;
+  std::cout << "starting enviroment" << std::endl;
 
-    pcl::visualization::PCLVisualizer::Ptr viewer (new pcl::visualization::PCLVisualizer ("3D Viewer"));
-    CameraAngle setAngle = FPS;
-    initCamera(setAngle, viewer);
+  pcl::visualization::PCLVisualizer::Ptr viewer(new pcl::visualization::PCLVisualizer("3D Viewer"));
+  CameraAngle setAngle = FPS;
+  initCamera(setAngle, viewer);
 
-    // Stream PCD Data
-    ProcessPointClouds<pcl::PointXYZI>* pointProcessorI = new ProcessPointClouds<pcl::PointXYZI>();
+  // Stream PCD Data
+  ProcessPointClouds<pcl::PointXYZI>* pointProcessorI = new ProcessPointClouds<pcl::PointXYZI>();
 
-    std::vector<boost::filesystem::path> stream = pointProcessorI->streamPcd("../src/sensors/data/pcd/data_1");
-    auto streamIterator = stream.begin();
+  std::vector<boost::filesystem::path> stream = pointProcessorI->streamPcd("../src/sensors/data/pcd/data_1");
+  auto streamIterator = stream.begin();
 
-    pcl::PointCloud<pcl::PointXYZI>::Ptr inputCloudI;
+  pcl::PointCloud<pcl::PointXYZI>::Ptr inputCloudI;
 
-    //simpleHighway(viewer);
-    //cityBlock(viewer, pointProcessorI, inputCloudI);
+  // simpleHighway(viewer);
+  // cityBlock(viewer, pointProcessorI, inputCloudI);
 
-    
-    while (!viewer->wasStopped ())
+  while (!viewer->wasStopped())
+  {
+    // Clear Viewer
+    viewer->removeAllPointClouds();
+    viewer->removeAllShapes();
+
+    // Load pcd and run obstacle detection process
+    inputCloudI = pointProcessorI->loadPcd((*streamIterator).string());
+    cityBlock(viewer, pointProcessorI, inputCloudI);
+
+    streamIterator++;
+    if (streamIterator == stream.end())
     {
-        // Clear Viewer
-        viewer->removeAllPointClouds();
-        viewer->removeAllShapes();
+      streamIterator = stream.begin();
+    }
 
-        // Load pcd and run obstacle detection process
-        inputCloudI = pointProcessorI->loadPcd((*streamIterator).string());
-        cityBlock(viewer, pointProcessorI, inputCloudI);
-
-        streamIterator++;
-        if(streamIterator == stream.end())
-            streamIterator = stream.begin();
-
-
-        viewer->spinOnce ();
-    } 
+    viewer->spinOnce();
+  }
 }
