@@ -1,7 +1,14 @@
-// PCL lib Functions for processing point clouds 
+// PCL lib Functions for processing point clouds
 
 #ifndef PROCESSPOINTCLOUDS_H_
 #define PROCESSPOINTCLOUDS_H_
+
+#include <iostream>
+#include <string>
+#include <vector>
+#include <ctime>
+#include <chrono>
+#include <unordered_set>
 
 #include <pcl/io/pcd_io.h>
 #include <pcl/common/common.h>
@@ -12,51 +19,53 @@
 #include <pcl/segmentation/sac_segmentation.h>
 #include <pcl/segmentation/extract_clusters.h>
 #include <pcl/common/transforms.h>
-#include <iostream> 
-#include <string>  
-#include <vector>
-#include <ctime>
-#include <chrono>
+#include <boost/filesystem.hpp>
+
+#include "quiz/cluster/kdtree.h"
 #include "render/box.h"
-#include <unordered_set>
-#include "./quiz/cluster/kdtree.h"
 
-template<typename PointT>
-class ProcessPointClouds {
+template <typename PointT>
+class ProcessPointClouds
+{
 public:
+  // constructor
+  ProcessPointClouds();
+  // deconstructor
+  ~ProcessPointClouds();
 
-    //constructor
-    ProcessPointClouds();
-    //deconstructor
-    ~ProcessPointClouds();
+  void numPoints(typename pcl::PointCloud<PointT>::Ptr cloud);
 
-    void numPoints(typename pcl::PointCloud<PointT>::Ptr cloud);
+  typename pcl::PointCloud<PointT>::Ptr FilterCloud(typename pcl::PointCloud<PointT>::Ptr cloud, float filterRes,
+                                                    Eigen::Vector4f minPoint, Eigen::Vector4f maxPoint);
 
-    typename pcl::PointCloud<PointT>::Ptr FilterCloud(typename pcl::PointCloud<PointT>::Ptr cloud, float filterRes, Eigen::Vector4f minPoint, Eigen::Vector4f maxPoint);
+  std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr>
+  SeparateClouds(pcl::PointIndices::Ptr inliers, typename pcl::PointCloud<PointT>::Ptr cloud);
 
-    std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr> SeparateClouds(pcl::PointIndices::Ptr inliers, typename pcl::PointCloud<PointT>::Ptr cloud);
+  std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr>
+  SegmentPlane(typename pcl::PointCloud<PointT>::Ptr cloud, int maxIterations, float distanceThreshold);
 
-    std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr> SegmentPlane(typename pcl::PointCloud<PointT>::Ptr cloud, int maxIterations, float distanceThreshold);
+  std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr>
+  RansacPlane(typename pcl::PointCloud<PointT>::Ptr cloud, int maxIterations, float distanceThreshold);
 
-    std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr> RansacPlane(typename pcl::PointCloud<PointT>::Ptr cloud, int maxIterations, float distanceThreshold);
+  std::vector<typename pcl::PointCloud<PointT>::Ptr> Clustering(typename pcl::PointCloud<PointT>::Ptr cloud,
+                                                                float clusterTolerance, int minSize, int maxSize);
 
-    std::vector<typename pcl::PointCloud<PointT>::Ptr> Clustering(typename pcl::PointCloud<PointT>::Ptr cloud, float clusterTolerance, int minSize, int maxSize);
+  void clusterHelper(int indice, const std::vector<std::vector<float>>& points, std::vector<int>& cluster,
+                     std::vector<bool>& processed, KdTree* tree, float distanceTol);
 
-    void clusterHelper(int indice, const std::vector<std::vector<float>>& points, std::vector<int>& cluster, std::vector<bool>& processed, KdTree * tree, float distanceTol );
+  std::vector<std::vector<int>> euclideanCluster(const std::vector<std::vector<float>>& points, KdTree* tree,
+                                                 float distanceTol);
 
-    std::vector<std::vector<int>> euclideanCluster(const std::vector<std::vector<float>>& points, KdTree* tree, float distanceTol);
+  std::vector<typename pcl::PointCloud<PointT>::Ptr> Clustering_Custom(typename pcl::PointCloud<PointT>::Ptr cloud,
+                                                                       float clusterTolerance, int minSize,
+                                                                       int maxSize);
 
-    std::vector<typename pcl::PointCloud<PointT>::Ptr> Clustering_Custom(typename pcl::PointCloud<PointT>::Ptr cloud, float clusterTolerance, int minSize, int maxSize);
+  Box BoundingBox(typename pcl::PointCloud<PointT>::Ptr cluster);
 
-    Box BoundingBox(typename pcl::PointCloud<PointT>::Ptr cluster);
+  void savePcd(typename pcl::PointCloud<PointT>::Ptr cloud, std::string file);
 
-    void savePcd(typename pcl::PointCloud<PointT>::Ptr cloud, std::string file);
+  typename pcl::PointCloud<PointT>::Ptr loadPcd(std::string file);
 
-    typename pcl::PointCloud<PointT>::Ptr loadPcd(std::string file);
-
-    std::vector<boost::filesystem::path> streamPcd(std::string dataPath);
-
-    
-  
+  std::vector<boost::filesystem::path> streamPcd(std::string dataPath);
 };
 #endif /* PROCESSPOINTCLOUDS_H_ */
