@@ -1,414 +1,409 @@
-// PCL lib Functions for processing point clouds 
+// PCL lib Functions for processing point clouds
 
 #include "processPointClouds.h"
 
-//constructor:
-template<typename PointT>
-ProcessPointClouds<PointT>::ProcessPointClouds() {}
-
-
-//de-constructor:
-template<typename PointT>
-ProcessPointClouds<PointT>::~ProcessPointClouds() {}
-
-
-template<typename PointT>
-void ProcessPointClouds<PointT>::numPoints(typename pcl::PointCloud<PointT>::Ptr cloud)
+// constructor:
+template <typename PointT>
+ProcessPointClouds<PointT>::ProcessPointClouds()
 {
-    std::cout << cloud->points.size() << std::endl;
 }
 
-
-template<typename PointT>
-typename pcl::PointCloud<PointT>::Ptr ProcessPointClouds<PointT>::FilterCloud(typename pcl::PointCloud<PointT>::Ptr cloud, float filterRes, Eigen::Vector4f minPoint, Eigen::Vector4f maxPoint)
+// de-constructor:
+template <typename PointT>
+ProcessPointClouds<PointT>::~ProcessPointClouds()
 {
-
-    // Time segmentation process
-    auto startTime = std::chrono::steady_clock::now();
-
-    // TODO:: Fill in the function to do voxel grid point reduction and region based filtering
-    // Create the filtering object: downsample the dataset using a leaf size of 0.2m
-    pcl::VoxelGrid<PointT> vg;
-    typename pcl::PointCloud<PointT>::Ptr cloudFiltered (new pcl::PointCloud<PointT>);
-    // std::cout << typeid(vg).name() << endl;
-    vg.setInputCloud(cloud);
-    vg.setLeafSize(filterRes, filterRes, filterRes);
-    vg.filter(*cloudFiltered);
-
-    typename pcl::PointCloud<PointT>::Ptr cloudRegion (new pcl::PointCloud<PointT>);
-
-    pcl::CropBox<PointT> region(true);
-    region.setMin(minPoint);
-    region.setMax(maxPoint);
-    region.setInputCloud(cloudFiltered);
-    region.filter(*cloudRegion);
-
-    std::vector<int> indices;
-
-    pcl::CropBox<PointT> roof(true);
-    roof.setMin(Eigen::Vector4f (-1.5, -1.7, -1, 1));
-    roof.setMax(Eigen::Vector4f (2.6, 1.7, -0.4, 1));
-    roof.setInputCloud(cloudRegion);
-    roof.filter(indices);
-
-    pcl::PointIndices::Ptr inliers {new pcl::PointIndices};
-    for(int point : indices)
-        inliers->indices.push_back(point);
-
-    pcl::ExtractIndices<PointT> extract;
-    extract.setInputCloud (cloudRegion);
-    extract.setIndices (inliers);
-    extract.setNegative (true);
-    extract.filter (*cloudRegion);
-
-    auto endTime = std::chrono::steady_clock::now();
-    auto elapsedTime = std::chrono::duration_cast<std::chrono::milliseconds>(endTime - startTime);
-    std::cout << "filtering took " << elapsedTime.count() << " milliseconds" << std::endl;
-
-    return cloudRegion;
-
-}
-
-
-template<typename PointT>
-std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr> ProcessPointClouds<PointT>::SeparateClouds(pcl::PointIndices::Ptr inliers, typename pcl::PointCloud<PointT>::Ptr cloud) 
-{
-  // TODO: Create two new point clouds, one cloud with obstacles and other with segmented plane
-    typename pcl::PointCloud<PointT>::Ptr obstCloud (new pcl::PointCloud<PointT> ());
-    typename pcl::PointCloud<PointT>::Ptr planeCloud (new pcl::PointCloud<PointT> ());
-
-    for (int index : inliers->indices)
-        planeCloud->points.push_back(cloud->points[index]);
-    
-    pcl::ExtractIndices<PointT> extract;
-    extract.setInputCloud (cloud);
-    extract.setIndices (inliers);
-    extract.setNegative (true);
-    extract.filter (*obstCloud);
-
-    std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr> segResult(obstCloud, planeCloud);
-    return segResult;
-}
-
-
-template<typename PointT>
-std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr> ProcessPointClouds<PointT>::SegmentPlane(typename pcl::PointCloud<PointT>::Ptr cloud, int maxIterations, float distanceThreshold)
-{
-    // Time segmentation process
-    auto startTime = std::chrono::steady_clock::now();
-
-    // TODO:: Fill in this function to find inliers for the cloud.
-    pcl::SACSegmentation<PointT> seg;
-    pcl::PointIndices::Ptr inliers {new pcl::PointIndices};
-    pcl::ModelCoefficients::Ptr coefficients {new pcl::ModelCoefficients};
-    
-    seg.setOptimizeCoefficients(true);
-    seg.setModelType(pcl::SACMODEL_PLANE);
-    seg.setMethodType(pcl::SAC_RANSAC);
-    seg.setMaxIterations(maxIterations);
-    seg.setDistanceThreshold(distanceThreshold);
-
-    //Segment the largest planar component from the input cloud
-    seg.setInputCloud(cloud);
-    seg.segment (*inliers, *coefficients);
-    
-    if(inliers->indices.size() == 0){
-        std::cout << "Could not estimate a planar model for the given dataset." << std::endl;
-    }
-    std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr> segResult = SeparateClouds(inliers,cloud);
-
-    auto endTime = std::chrono::steady_clock::now();
-    auto elapsedTime = std::chrono::duration_cast<std::chrono::milliseconds>(endTime - startTime);
-    std::cout << "plane segmentation took " << elapsedTime.count() << " milliseconds" << std::endl;
-
-    return segResult;
-}
-
-template<typename PointT>
-std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr> ProcessPointClouds<PointT>::RansacPlane(typename pcl::PointCloud<PointT>::Ptr cloud, int maxIterations, float distanceTol)
-{
-	std::unordered_set<int> inliersResult;
-	srand(time(NULL));
-
-	// For max iterations 
-	// Randomly sample subset and fit line
-	// Measure distance between every point and fitted line
-	// If distance is smaller than threshold count it as inlier
-	// Return indicies of inliers from fitted line with most inliers
-	while(maxIterations--)
-	{
-
-		// Select two points randomly
-		std::unordered_set<int> inliers; // hash set
-		while(inliers.size() != 3)
-			inliers.insert(rand()%(cloud->points.size()));
-		
-		float x1, x2, x3, y1, y2, y3, z1, z2, z3;
-
-		auto itr = inliers.begin();
-		x1 = cloud->points[*itr].x;
-		y1 = cloud->points[*itr].y;
-		z1 = cloud->points[*itr].z;
-		itr++;
-		x2 = cloud->points[*itr].x;
-		y2 = cloud->points[*itr].y;
-		z2 = cloud->points[*itr].z;
-		itr++;
-		x3 = cloud->points[*itr].x;
-		y3 = cloud->points[*itr].y;
-		z3 = cloud->points[*itr].z;		
-
-		std::vector<float> v1 = {x2-x1, y2-y1, z2-z1};
-		std::vector<float> v2 = {x3-x1, y3-y1, z3-z1};
-
-		std::vector<float> cross_v1v2 = {v1[1]*v2[2]-v1[2]*v2[1], v1[2]*v2[0]-v1[0]*v2[2], v1[0]*v2[1]-v1[1]*v2[0]};
-
-		float a = cross_v1v2[0];
-		float b = cross_v1v2[1];
-		float c = cross_v1v2[2];
-		float d = -(cross_v1v2[0]*x1 + cross_v1v2[1]*y1+cross_v1v2[2]*z1);
-
-		for(int index = 0; index < cloud->points.size(); index++)
-		{
-            if (inliers.count(index) > 0)
-                continue;
-
-			PointT point = cloud->points[index];
-			float x4 = point.x;
-			float y4 = point.y;
-			float z4 = point.z;
-
-			float dist = fabs(a*x4+b*y4+c*z4+d)/sqrt(a*a+b*b+c*c);
-
-			if(dist <= distanceTol)
-				inliers.insert(index);
-
-		}
-
-		if(inliers.size()>inliersResult.size())
-		{
-			inliersResult = inliers;
-		}
-        
-	}
-
-    //Convert unordered_set into pairs
-    typename pcl::PointCloud<PointT>::Ptr  cloudInliers(new pcl::PointCloud<PointT>());
-    typename pcl::PointCloud<PointT>::Ptr cloudOutliers(new pcl::PointCloud<PointT>());
-
-    for(int i = 0; i < cloud->points.size(); i++)
-    {
-        PointT point = cloud->points[i];
-        if(inliersResult.count(i))
-            cloudInliers->points.push_back(point);
-        else
-            cloudOutliers->points.push_back(point);
-    }
-
-    std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr> segResult(cloudOutliers, cloudInliers);
-
-
-
-	return segResult;
-}
-
-
-template<typename PointT>
-std::vector<typename pcl::PointCloud<PointT>::Ptr> ProcessPointClouds<PointT>::Clustering(typename pcl::PointCloud<PointT>::Ptr cloud, float clusterTolerance, int minSize, int maxSize)
-{
-
-    // Time clustering process
-    auto startTime = std::chrono::steady_clock::now();
-    
-    std::vector<typename pcl::PointCloud<PointT>::Ptr> clusters;
-    // TODO:: Fill in the function to perform euclidean clustering to group detected obstacles
-    // Create KDTree Object for the search method of the extraction
-    typename pcl::search::KdTree<PointT>::Ptr tree (new pcl::search::KdTree<PointT>);
-    tree->setInputCloud(cloud);
-
-    std::vector<pcl::PointIndices> clusterIndices;
-    pcl::EuclideanClusterExtraction<PointT> ec;
-    ec.setClusterTolerance(clusterTolerance);
-    ec.setMinClusterSize(minSize);
-    ec.setMaxClusterSize(maxSize);
-    ec.setSearchMethod(tree);
-    ec.setInputCloud(cloud);
-    ec.extract(clusterIndices);
-
-    for(pcl::PointIndices getIndices: clusterIndices)
-    {
-        typename pcl::PointCloud<PointT>::Ptr cloudCluster (new pcl::PointCloud<PointT>);
-
-        for(int index : getIndices.indices)
-            cloudCluster->points.push_back (cloud->points[index]);
-
-        cloudCluster->width = cloudCluster->points.size();
-        cloudCluster->height = 1;
-        cloudCluster->is_dense = true;
-
-        clusters.push_back(cloudCluster);
-    }
-    
-    auto endTime = std::chrono::steady_clock::now();
-    auto elapsedTime = std::chrono::duration_cast<std::chrono::milliseconds>(endTime - startTime);
-    std::cout << "clustering took " << elapsedTime.count() << " milliseconds and found " << clusters.size() << " clusters" << std::endl;
-    
-    return clusters;
-}
-
-template<typename PointT>
-void ProcessPointClouds<PointT>::clusterHelper(int indice, const std::vector<std::vector<float>>& points, std::vector<int>& cluster, std::vector<bool>& processed, KdTree * tree, float distanceTol )
-{
-	processed[indice] = true;
-	cluster.push_back(indice);
-
-	std::vector<int> nearest = tree->search(points[indice], distanceTol);
-
-	for(int id : nearest)
-	{
-		if(!processed[id])
-			clusterHelper(id, points, cluster, processed, tree, distanceTol);
-	}
 }
 
 template <typename PointT>
-std::vector<std::vector<int>> ProcessPointClouds<PointT>::euclideanCluster(const std::vector<std::vector<float>>& points, KdTree* tree, float distanceTol)
+void ProcessPointClouds<PointT>::numPoints(typename pcl::PointCloud<PointT>::Ptr cloud)
 {
-
-	// TODO: Fill out this function to return list of indices for each cluster
-
-	std::vector<std::vector<int>> clusters; // matrix of clusters
-
-	std::vector<bool> processed(points.size(), false);
- 
-	int i = 0; //initialize a counter
-	while(i < points.size())
-	{
-		if(processed[i]) // if the point has been already processed, skip it
-		{
-			i++;
-			continue;
-		}
-
-		std::vector<int> cluster; // create new cluster for the point
-		clusterHelper(i, points, cluster, processed, tree, distanceTol);
-		clusters.push_back(cluster);
-		i++;
-	}
-
-
-
-	return clusters;
-
+  std::cout << cloud->points.size() << std::endl;
 }
 
-template<typename PointT>
-std::vector<typename pcl::PointCloud<PointT>::Ptr> ProcessPointClouds<PointT>::Clustering_Custom(typename pcl::PointCloud<PointT>::Ptr cloud, float clusterTolerance, int minSize, int maxSize)
+template <typename PointT>
+typename pcl::PointCloud<PointT>::Ptr ProcessPointClouds<PointT>::FilterCloud(
+    typename pcl::PointCloud<PointT>::Ptr cloud, float filterRes, Eigen::Vector4f minPoint, Eigen::Vector4f maxPoint)
 {
+  // Time segmentation process
+  auto startTime = std::chrono::steady_clock::now();
 
-    // Time clustering process
-    auto startTime = std::chrono::steady_clock::now();
+  // TODO:: Fill in the function to do voxel grid point reduction and region based filtering
+  // Create the filtering object: downsample the dataset using a leaf size of 0.2m
+  pcl::VoxelGrid<PointT> vg;
+  typename pcl::PointCloud<PointT>::Ptr cloudFiltered(new pcl::PointCloud<PointT>);
+  // std::cout << typeid(vg).name() << endl;
+  vg.setInputCloud(cloud);
+  vg.setLeafSize(filterRes, filterRes, filterRes);
+  vg.filter(*cloudFiltered);
 
-    // Setup KdTree
-    KdTree* tree = new KdTree;
-    std::vector<std::vector<float>> points;
-    
-    int i = 0;
-    for (auto point : cloud->points)
-    {
-        const std::vector<float> p{point.x, point.y, point.z};
-        tree->insert(p, i++);
-        points.push_back(p);
-    }
+  typename pcl::PointCloud<PointT>::Ptr cloudRegion(new pcl::PointCloud<PointT>);
 
-    // Setup the Cluster
-    std::vector<std::vector<int>> clusters_indices = euclideanCluster(points, tree, clusterTolerance);
+  pcl::CropBox<PointT> region(true);
+  region.setMin(minPoint);
+  region.setMax(maxPoint);
+  region.setInputCloud(cloudFiltered);
+  region.filter(*cloudRegion);
 
-    std::vector<typename pcl::PointCloud<PointT>::Ptr> clusters;
+  std::vector<int> indices;
 
-    for(auto indices : clusters_indices)
-    {
-        if (indices.size() < minSize || indices.size() > maxSize) {continue;}
-        
-        typename pcl::PointCloud<PointT>::Ptr clusterCloud(new pcl::PointCloud<PointT>());
-        
-        for(auto index : indices)
-        {
-            clusterCloud->points.push_back(cloud->points[index]);
-        }
-        
-        clusterCloud->width = clusterCloud->points.size();
-        clusterCloud->height = 1;
-        clusterCloud->is_dense = true;
-        
-        clusters.push_back(clusterCloud);
-    }
-    
-    auto endTime = std::chrono::steady_clock::now();
-    auto elapsedTime = std::chrono::duration_cast<std::chrono::milliseconds>(endTime - startTime);
-    std::cout << "clustering took " << elapsedTime.count() << " milliseconds and found " << clusters.size() << " clusters" << std::endl;
+  pcl::CropBox<PointT> roof(true);
+  roof.setMin(Eigen::Vector4f(-1.5, -1.7, -1, 1));
+  roof.setMax(Eigen::Vector4f(2.6, 1.7, -0.4, 1));
+  roof.setInputCloud(cloudRegion);
+  roof.filter(indices);
 
-    return clusters;
+  pcl::PointIndices::Ptr inliers{ new pcl::PointIndices };
+  for (int point : indices)
+    inliers->indices.push_back(point);
+
+  pcl::ExtractIndices<PointT> extract;
+  extract.setInputCloud(cloudRegion);
+  extract.setIndices(inliers);
+  extract.setNegative(true);
+  extract.filter(*cloudRegion);
+
+  auto endTime = std::chrono::steady_clock::now();
+  auto elapsedTime = std::chrono::duration_cast<std::chrono::milliseconds>(endTime - startTime);
+  std::cout << "filtering took " << elapsedTime.count() << " milliseconds" << std::endl;
+
+  return cloudRegion;
 }
 
+template <typename PointT>
+std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr>
+ProcessPointClouds<PointT>::SeparateClouds(pcl::PointIndices::Ptr inliers, typename pcl::PointCloud<PointT>::Ptr cloud)
+{
+  // TODO: Create two new point clouds, one cloud with obstacles and other with segmented plane
+  typename pcl::PointCloud<PointT>::Ptr obstCloud(new pcl::PointCloud<PointT>());
+  typename pcl::PointCloud<PointT>::Ptr planeCloud(new pcl::PointCloud<PointT>());
 
-template<typename PointT>
+  for (int index : inliers->indices)
+    planeCloud->points.push_back(cloud->points[index]);
+
+  pcl::ExtractIndices<PointT> extract;
+  extract.setInputCloud(cloud);
+  extract.setIndices(inliers);
+  extract.setNegative(true);
+  extract.filter(*obstCloud);
+
+  std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr> segResult(obstCloud,
+                                                                                                    planeCloud);
+  return segResult;
+}
+
+template <typename PointT>
+std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr>
+ProcessPointClouds<PointT>::SegmentPlane(typename pcl::PointCloud<PointT>::Ptr cloud, int maxIterations,
+                                         float distanceThreshold)
+{
+  // Time segmentation process
+  auto startTime = std::chrono::steady_clock::now();
+
+  // TODO:: Fill in this function to find inliers for the cloud.
+  pcl::SACSegmentation<PointT> seg;
+  pcl::PointIndices::Ptr inliers{ new pcl::PointIndices };
+  pcl::ModelCoefficients::Ptr coefficients{ new pcl::ModelCoefficients };
+
+  seg.setOptimizeCoefficients(true);
+  seg.setModelType(pcl::SACMODEL_PLANE);
+  seg.setMethodType(pcl::SAC_RANSAC);
+  seg.setMaxIterations(maxIterations);
+  seg.setDistanceThreshold(distanceThreshold);
+
+  // Segment the largest planar component from the input cloud
+  seg.setInputCloud(cloud);
+  seg.segment(*inliers, *coefficients);
+
+  if (inliers->indices.size() == 0)
+  {
+    std::cout << "Could not estimate a planar model for the given dataset." << std::endl;
+  }
+  std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr> segResult =
+      SeparateClouds(inliers, cloud);
+
+  auto endTime = std::chrono::steady_clock::now();
+  auto elapsedTime = std::chrono::duration_cast<std::chrono::milliseconds>(endTime - startTime);
+  std::cout << "plane segmentation took " << elapsedTime.count() << " milliseconds" << std::endl;
+
+  return segResult;
+}
+
+template <typename PointT>
+std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr>
+ProcessPointClouds<PointT>::RansacPlane(typename pcl::PointCloud<PointT>::Ptr cloud, int maxIterations,
+                                        float distanceTol)
+{
+  std::unordered_set<int> inliersResult;
+  srand(time(NULL));
+
+  // For max iterations
+  // Randomly sample subset and fit line
+  // Measure distance between every point and fitted line
+  // If distance is smaller than threshold count it as inlier
+  // Return indicies of inliers from fitted line with most inliers
+  while (maxIterations--)
+  {
+    // Select two points randomly
+    std::unordered_set<int> inliers;  // hash set
+    while (inliers.size() != 3)
+      inliers.insert(rand() % (cloud->points.size()));
+
+    float x1, x2, x3, y1, y2, y3, z1, z2, z3;
+
+    auto itr = inliers.begin();
+    x1 = cloud->points[*itr].x;
+    y1 = cloud->points[*itr].y;
+    z1 = cloud->points[*itr].z;
+    itr++;
+    x2 = cloud->points[*itr].x;
+    y2 = cloud->points[*itr].y;
+    z2 = cloud->points[*itr].z;
+    itr++;
+    x3 = cloud->points[*itr].x;
+    y3 = cloud->points[*itr].y;
+    z3 = cloud->points[*itr].z;
+
+    std::vector<float> v1 = { x2 - x1, y2 - y1, z2 - z1 };
+    std::vector<float> v2 = { x3 - x1, y3 - y1, z3 - z1 };
+
+    std::vector<float> cross_v1v2 = { v1[1] * v2[2] - v1[2] * v2[1], v1[2] * v2[0] - v1[0] * v2[2],
+                                      v1[0] * v2[1] - v1[1] * v2[0] };
+
+    float a = cross_v1v2[0];
+    float b = cross_v1v2[1];
+    float c = cross_v1v2[2];
+    float d = -(cross_v1v2[0] * x1 + cross_v1v2[1] * y1 + cross_v1v2[2] * z1);
+
+    for (int index = 0; index < cloud->points.size(); index++)
+    {
+      if (inliers.count(index) > 0)
+        continue;
+
+      PointT point = cloud->points[index];
+      float x4 = point.x;
+      float y4 = point.y;
+      float z4 = point.z;
+
+      float dist = fabs(a * x4 + b * y4 + c * z4 + d) / sqrt(a * a + b * b + c * c);
+
+      if (dist <= distanceTol)
+        inliers.insert(index);
+    }
+
+    if (inliers.size() > inliersResult.size())
+    {
+      inliersResult = inliers;
+    }
+  }
+
+  // Convert unordered_set into pairs
+  typename pcl::PointCloud<PointT>::Ptr cloudInliers(new pcl::PointCloud<PointT>());
+  typename pcl::PointCloud<PointT>::Ptr cloudOutliers(new pcl::PointCloud<PointT>());
+
+  for (int i = 0; i < cloud->points.size(); i++)
+  {
+    PointT point = cloud->points[i];
+    if (inliersResult.count(i))
+      cloudInliers->points.push_back(point);
+    else
+      cloudOutliers->points.push_back(point);
+  }
+
+  std::pair<typename pcl::PointCloud<PointT>::Ptr, typename pcl::PointCloud<PointT>::Ptr> segResult(cloudOutliers,
+                                                                                                    cloudInliers);
+
+  return segResult;
+}
+
+template <typename PointT>
+std::vector<typename pcl::PointCloud<PointT>::Ptr> ProcessPointClouds<PointT>::Clustering(
+    typename pcl::PointCloud<PointT>::Ptr cloud, float clusterTolerance, int minSize, int maxSize)
+{
+  // Time clustering process
+  auto startTime = std::chrono::steady_clock::now();
+
+  std::vector<typename pcl::PointCloud<PointT>::Ptr> clusters;
+  // TODO:: Fill in the function to perform euclidean clustering to group detected obstacles
+  // Create KDTree Object for the search method of the extraction
+  typename pcl::search::KdTree<PointT>::Ptr tree(new pcl::search::KdTree<PointT>);
+  tree->setInputCloud(cloud);
+
+  std::vector<pcl::PointIndices> clusterIndices;
+  pcl::EuclideanClusterExtraction<PointT> ec;
+  ec.setClusterTolerance(clusterTolerance);
+  ec.setMinClusterSize(minSize);
+  ec.setMaxClusterSize(maxSize);
+  ec.setSearchMethod(tree);
+  ec.setInputCloud(cloud);
+  ec.extract(clusterIndices);
+
+  for (pcl::PointIndices getIndices : clusterIndices)
+  {
+    typename pcl::PointCloud<PointT>::Ptr cloudCluster(new pcl::PointCloud<PointT>);
+
+    for (int index : getIndices.indices)
+      cloudCluster->points.push_back(cloud->points[index]);
+
+    cloudCluster->width = cloudCluster->points.size();
+    cloudCluster->height = 1;
+    cloudCluster->is_dense = true;
+
+    clusters.push_back(cloudCluster);
+  }
+
+  auto endTime = std::chrono::steady_clock::now();
+  auto elapsedTime = std::chrono::duration_cast<std::chrono::milliseconds>(endTime - startTime);
+  std::cout << "clustering took " << elapsedTime.count() << " milliseconds and found " << clusters.size() << " clusters"
+            << std::endl;
+
+  return clusters;
+}
+
+template <typename PointT>
+void ProcessPointClouds<PointT>::clusterHelper(int indice, const std::vector<std::vector<float>>& points,
+                                               std::vector<int>& cluster, std::vector<bool>& processed, KdTree* tree,
+                                               float distanceTol)
+{
+  processed[indice] = true;
+  cluster.push_back(indice);
+
+  std::vector<int> nearest = tree->search(points[indice], distanceTol);
+
+  for (int id : nearest)
+  {
+    if (!processed[id])
+      clusterHelper(id, points, cluster, processed, tree, distanceTol);
+  }
+}
+
+template <typename PointT>
+std::vector<std::vector<int>> ProcessPointClouds<PointT>::euclideanCluster(
+    const std::vector<std::vector<float>>& points, KdTree* tree, float distanceTol)
+{
+  // TODO: Fill out this function to return list of indices for each cluster
+
+  std::vector<std::vector<int>> clusters;  // matrix of clusters
+
+  std::vector<bool> processed(points.size(), false);
+
+  int i = 0;  // initialize a counter
+  while (i < points.size())
+  {
+    if (processed[i])  // if the point has been already processed, skip it
+    {
+      i++;
+      continue;
+    }
+
+    std::vector<int> cluster;  // create new cluster for the point
+    clusterHelper(i, points, cluster, processed, tree, distanceTol);
+    clusters.push_back(cluster);
+    i++;
+  }
+
+  return clusters;
+}
+
+template <typename PointT>
+std::vector<typename pcl::PointCloud<PointT>::Ptr> ProcessPointClouds<PointT>::Clustering_Custom(
+    typename pcl::PointCloud<PointT>::Ptr cloud, float clusterTolerance, int minSize, int maxSize)
+{
+  // Time clustering process
+  auto startTime = std::chrono::steady_clock::now();
+
+  // Setup KdTree
+  KdTree* tree = new KdTree;
+  std::vector<std::vector<float>> points;
+
+  int i = 0;
+  for (auto point : cloud->points)
+  {
+    const std::vector<float> p{ point.x, point.y, point.z };
+    tree->insert(p, i++);
+    points.push_back(p);
+  }
+
+  // Setup the Cluster
+  std::vector<std::vector<int>> clusters_indices = euclideanCluster(points, tree, clusterTolerance);
+
+  std::vector<typename pcl::PointCloud<PointT>::Ptr> clusters;
+
+  for (auto indices : clusters_indices)
+  {
+    if (indices.size() < minSize || indices.size() > maxSize)
+    {
+      continue;
+    }
+
+    typename pcl::PointCloud<PointT>::Ptr clusterCloud(new pcl::PointCloud<PointT>());
+
+    for (auto index : indices)
+    {
+      clusterCloud->points.push_back(cloud->points[index]);
+    }
+
+    clusterCloud->width = clusterCloud->points.size();
+    clusterCloud->height = 1;
+    clusterCloud->is_dense = true;
+
+    clusters.push_back(clusterCloud);
+  }
+
+  auto endTime = std::chrono::steady_clock::now();
+  auto elapsedTime = std::chrono::duration_cast<std::chrono::milliseconds>(endTime - startTime);
+  std::cout << "clustering took " << elapsedTime.count() << " milliseconds and found " << clusters.size() << " clusters"
+            << std::endl;
+
+  return clusters;
+}
+
+template <typename PointT>
 Box ProcessPointClouds<PointT>::BoundingBox(typename pcl::PointCloud<PointT>::Ptr cluster)
 {
+  // Find bounding box for one of the clusters
+  PointT minPoint, maxPoint;
+  pcl::getMinMax3D(*cluster, minPoint, maxPoint);
 
-    // Find bounding box for one of the clusters
-    PointT minPoint, maxPoint;
-    pcl::getMinMax3D(*cluster, minPoint, maxPoint);
+  Box box;
+  box.x_min = minPoint.x;
+  box.y_min = minPoint.y;
+  box.z_min = minPoint.z;
+  box.x_max = maxPoint.x;
+  box.y_max = maxPoint.y;
+  box.z_max = maxPoint.z;
 
-    Box box;
-    box.x_min = minPoint.x;
-    box.y_min = minPoint.y;
-    box.z_min = minPoint.z;
-    box.x_max = maxPoint.x;
-    box.y_max = maxPoint.y;
-    box.z_max = maxPoint.z;
-
-    return box;
+  return box;
 }
 
-
-template<typename PointT>
+template <typename PointT>
 void ProcessPointClouds<PointT>::savePcd(typename pcl::PointCloud<PointT>::Ptr cloud, std::string file)
 {
-    pcl::io::savePCDFileASCII (file, *cloud);
-    std::cerr << "Saved " << cloud->points.size () << " data points to "+file << std::endl;
+  pcl::io::savePCDFileASCII(file, *cloud);
+  std::cerr << "Saved " << cloud->points.size() << " data points to " + file << std::endl;
 }
 
-
-template<typename PointT>
+template <typename PointT>
 typename pcl::PointCloud<PointT>::Ptr ProcessPointClouds<PointT>::loadPcd(std::string file)
 {
+  typename pcl::PointCloud<PointT>::Ptr cloud(new pcl::PointCloud<PointT>);
 
-    typename pcl::PointCloud<PointT>::Ptr cloud (new pcl::PointCloud<PointT>);
+  if (pcl::io::loadPCDFile<PointT>(file, *cloud) == -1)  //* load the file
+  {
+    PCL_ERROR("Couldn't read file \n");
+  }
+  std::cerr << "Loaded " << cloud->points.size() << " data points from " + file << std::endl;
 
-    if (pcl::io::loadPCDFile<PointT> (file, *cloud) == -1) //* load the file
-    {
-        PCL_ERROR ("Couldn't read file \n");
-    }
-    std::cerr << "Loaded " << cloud->points.size () << " data points from "+file << std::endl;
-
-    return cloud;
+  return cloud;
 }
 
-
-template<typename PointT>
+template <typename PointT>
 std::vector<boost::filesystem::path> ProcessPointClouds<PointT>::streamPcd(std::string dataPath)
 {
+  std::vector<boost::filesystem::path> paths(boost::filesystem::directory_iterator{ dataPath },
+                                             boost::filesystem::directory_iterator{});
 
-    std::vector<boost::filesystem::path> paths(boost::filesystem::directory_iterator{dataPath}, boost::filesystem::directory_iterator{});
+  // sort files in accending order so playback is chronological
+  sort(paths.begin(), paths.end());
 
-    // sort files in accending order so playback is chronological
-    sort(paths.begin(), paths.end());
-
-    return paths;
-
+  return paths;
 }
-
-
-
-
