@@ -1,47 +1,46 @@
 /* \author Aaron Brown */
 // Quiz on implementing simple RANSAC line fitting
-
-#include "../../render/render.h"
 #include <unordered_set>
-#include "../../processPointClouds.h"
+
+#include "render/render.h"
+
+#include "processPointClouds.h"
 // using templates for processPointClouds so also include .cpp to help linker
-#include "../../processPointClouds.cpp"
+#include "processPointClouds.cpp"
 
 pcl::PointCloud<pcl::PointXYZ>::Ptr CreateData()
 {
 	pcl::PointCloud<pcl::PointXYZ>::Ptr cloud(new pcl::PointCloud<pcl::PointXYZ>());
-  	// Add inliers
-  	float scatter = 0.6;
-  	for(int i = -5; i < 5; i++)
-  	{
-  		double rx = 2*(((double) rand() / (RAND_MAX))-0.5);
-  		double ry = 2*(((double) rand() / (RAND_MAX))-0.5);
-  		pcl::PointXYZ point;
-  		point.x = i+scatter*rx;
-  		point.y = i+scatter*ry;
-  		point.z = 0;
+	// Add inliers
+	float scatter = 0.6;
+	for (int i = -5; i < 5; i++)
+	{
+		double rx = 2 * (((double)rand() / (RAND_MAX)) - 0.5);
+		double ry = 2 * (((double)rand() / (RAND_MAX)) - 0.5);
+		pcl::PointXYZ point;
+		point.x = i + scatter * rx;
+		point.y = i + scatter * ry;
+		point.z = 0;
 
-  		cloud->points.push_back(point);
-  	}
-  	// Add outliers
-  	int numOutliers = 10;
-  	while(numOutliers--)
-  	{
-  		double rx = 2*(((double) rand() / (RAND_MAX))-0.5);
-  		double ry = 2*(((double) rand() / (RAND_MAX))-0.5);
-  		pcl::PointXYZ point;
-  		point.x = 5*rx;
-  		point.y = 5*ry;
-  		point.z = 0;
+		cloud->points.push_back(point);
+	}
+	// Add outliers
+	int numOutliers = 10;
+	while (numOutliers--)
+	{
+		double rx = 2 * (((double)rand() / (RAND_MAX)) - 0.5);
+		double ry = 2 * (((double)rand() / (RAND_MAX)) - 0.5);
+		pcl::PointXYZ point;
+		point.x = 5 * rx;
+		point.y = 5 * ry;
+		point.z = 0;
 
-  		cloud->points.push_back(point);
+		cloud->points.push_back(point);
+	}
+	cloud->width = cloud->points.size();
+	cloud->height = 1;
 
-  	}
-  	cloud->width = cloud->points.size();
-  	cloud->height = 1;
-
-  	return cloud;
-
+	return cloud;
 }
 
 pcl::PointCloud<pcl::PointXYZ>::Ptr CreateData3D()
@@ -50,25 +49,24 @@ pcl::PointCloud<pcl::PointXYZ>::Ptr CreateData3D()
 	return pointProcessor.loadPcd("../../../sensors/data/pcd/simpleHighway.pcd");
 }
 
-
 pcl::visualization::PCLVisualizer::Ptr initScene()
 {
-	pcl::visualization::PCLVisualizer::Ptr viewer(new pcl::visualization::PCLVisualizer ("2D Viewer"));
-	viewer->setBackgroundColor (0, 0, 0);
-  	viewer->initCameraParameters();
-  	viewer->setCameraPosition(0, 0, 15, 0, 1, 0);
-  	viewer->addCoordinateSystem (1.0);
-  	return viewer;
+	pcl::visualization::PCLVisualizer::Ptr viewer(new pcl::visualization::PCLVisualizer("2D Viewer"));
+	viewer->setBackgroundColor(0, 0, 0);
+	viewer->initCameraParameters();
+	viewer->setCameraPosition(0, 0, 15, 0, 1, 0);
+	viewer->addCoordinateSystem(1.0);
+	return viewer;
 }
 
 std::unordered_set<int> RansacPlane(pcl::PointCloud<pcl::PointXYZ>::Ptr cloud, int maxIterations, float distanceTol)
 {
 	std::unordered_set<int> inliersResult;
 	srand(time(NULL));
-	
+
 	// TODO: Fill in this function
 
-	// For max iterations 
+	// For max iterations
 
 	// Randomly sample subset and fit line
 
@@ -76,14 +74,13 @@ std::unordered_set<int> RansacPlane(pcl::PointCloud<pcl::PointXYZ>::Ptr cloud, i
 	// If distance is smaller than threshold count it as inlier
 
 	// Return indicies of inliers from fitted line with most inliers
-	while(maxIterations--)
+	while (maxIterations--)
 	{
-
 		// Select two points randomly
-		std::unordered_set<int> inliers; // hash set
-		while(inliers.size() < 3)
-			inliers.insert(rand()%(cloud->points.size()));
-		
+		std::unordered_set<int> inliers;	// hash set
+		while (inliers.size() < 3)
+			inliers.insert(rand() % (cloud->points.size()));
+
 		float x1, x2, x3, y1, y2, y3, z1, z2, z3;
 
 		auto itr = inliers.begin();
@@ -97,84 +94,79 @@ std::unordered_set<int> RansacPlane(pcl::PointCloud<pcl::PointXYZ>::Ptr cloud, i
 		itr++;
 		x3 = cloud->points[*itr].x;
 		y3 = cloud->points[*itr].y;
-		z3 = cloud->points[*itr].z;		
+		z3 = cloud->points[*itr].z;
 
-		std::vector<float> v1 = {x2-x1, y2-y1, z2-z1};
-		std::vector<float> v2 = {x3-x1, y3-y1, z3-z1};
+		std::vector<float> v1 = { x2 - x1, y2 - y1, z2 - z1 };
+		std::vector<float> v2 = { x3 - x1, y3 - y1, z3 - z1 };
 
-		std::vector<float> cross_v1v2 = {v1[1]*v2[2]-v1[2]*v2[1], v1[2]*v2[0]-v1[0]*v2[2], v1[0]*v2[1]-v1[1]*v2[0]};
+		std::vector<float> cross_v1v2 = { v1[1] * v2[2] - v1[2] * v2[1], v1[2] * v2[0] - v1[0] * v2[2],
+																			v1[0] * v2[1] - v1[1] * v2[0] };
 
 		float a = cross_v1v2[0];
 		float b = cross_v1v2[1];
 		float c = cross_v1v2[2];
-		float d = -(cross_v1v2[0]*x1 + cross_v1v2[1]*y1+cross_v1v2[2]*z1);
+		float d = -(cross_v1v2[0] * x1 + cross_v1v2[1] * y1 + cross_v1v2[2] * z1);
 
-		for(int index = 0; index < cloud->points.size(); index++)
+		for (int index = 0; index < cloud->points.size(); index++)
 		{
-			if(inliers.count(index)>0)
+			if (inliers.count(index) > 0)
 				continue;
-			
+
 			pcl::PointXYZ point = cloud->points[index];
 			float x4 = point.x;
 			float y4 = point.y;
 			float z4 = point.z;
 
-			float d = fabs(a*x4+b*y4+c*z4+d)/sqrt(a*a+b*b+c*c);
+			float d = fabs(a * x4 + b * y4 + c * z4 + d) / sqrt(a * a + b * b + c * c);
 
-			if(d <= distanceTol)
+			if (d <= distanceTol)
 				inliers.insert(index);
-
 		}
 
-		if(inliers.size()>inliersResult.size())
+		if (inliers.size() > inliersResult.size())
 		{
 			inliersResult = inliers;
 		}
 	}
 	return inliersResult;
-
 }
 
-int main ()
+int main()
 {
-
 	// Create viewer
 	pcl::visualization::PCLVisualizer::Ptr viewer = initScene();
 
 	// Create data
 	pcl::PointCloud<pcl::PointXYZ>::Ptr cloud = CreateData3D();
-	
 
 	// TODO: Change the max iteration and distance tolerance arguments for Ransac function
 	std::unordered_set<int> inliers = RansacPlane(cloud, 50, 0.3);
 
-	pcl::PointCloud<pcl::PointXYZ>::Ptr  cloudInliers(new pcl::PointCloud<pcl::PointXYZ>());
+	pcl::PointCloud<pcl::PointXYZ>::Ptr cloudInliers(new pcl::PointCloud<pcl::PointXYZ>());
 	pcl::PointCloud<pcl::PointXYZ>::Ptr cloudOutliers(new pcl::PointCloud<pcl::PointXYZ>());
 
-	for(int index = 0; index < cloud->points.size(); index++)
+	for (int index = 0; index < cloud->points.size(); index++)
 	{
 		pcl::PointXYZ point = cloud->points[index];
-		if(inliers.count(index))
+		if (inliers.count(index))
 			cloudInliers->points.push_back(point);
 		else
 			cloudOutliers->points.push_back(point);
 	}
 
-
 	// Render 2D point cloud with inliers and outliers
-	if(inliers.size())
+	if (inliers.size())
 	{
-		renderPointCloud(viewer,cloudInliers,"inliers",Color(0,1,0));
-  		renderPointCloud(viewer,cloudOutliers,"outliers",Color(1,0,0));
+		renderPointCloud(viewer, cloudInliers, "inliers", Color(0, 1, 0));
+		renderPointCloud(viewer, cloudOutliers, "outliers", Color(1, 0, 0));
 	}
-  	else
-  	{
-  		renderPointCloud(viewer,cloud,"data");
-  	}
-	
-  	while (!viewer->wasStopped ())
-  	{
-  	  viewer->spinOnce ();
-  	}
-  	
+	else
+	{
+		renderPointCloud(viewer, cloud, "data");
+	}
+
+	while (!viewer->wasStopped())
+	{
+		viewer->spinOnce();
+	}
 }
