@@ -1,15 +1,16 @@
 #ifndef BOX_H
 #define BOX_H
-#include <Eigen/Geometry> 
+#include <Eigen/Geometry>
 
 struct BoxQ
 {
 	Eigen::Vector3f bboxTransform;
 	Eigen::Quaternionf bboxQuaternion;
 	float cube_length;
-    float cube_width;
-    float cube_height;
+	float cube_width;
+	float cube_height;
 };
+
 struct Box
 {
 	float x_min;
