@@ -1,14 +1,15 @@
 #ifndef LIDAR_H
 #define LIDAR_H
-#include "../render/render.h"
-#include <ctime>
+
 #include <chrono>
+#include <ctime>
+
+#include "render/render.h"
 
 const double pi = 3.1415;
 
 struct Ray
 {
-	
 	Vect3 origin;
 	double resolution;
 	Vect3 direction;
@@ -18,16 +19,23 @@ struct Ray
 	// parameters:
 	// setOrigin: the starting position from where the ray is cast
 	// horizontalAngle: the angle of direction the ray travels on the xy plane
-	// verticalAngle: the angle of direction between xy plane and ray 
-	// 				  for example 0 radians is along xy plane and pi/2 radians is stright up
-	// resoultion: the magnitude of the ray's step, used for ray casting, the smaller the more accurate but the more expensive
+	// verticalAngle: the angle of direction between xy plane and ray
+	// 				  for example 0 radians is along xy plane and pi/2 radians is
+	// stright up resoultion: the magnitude of the ray's step, used for ray
+	// casting, the smaller the more accurate but the more expensive
 
 	Ray(Vect3 setOrigin, double horizontalAngle, double verticalAngle, double setResolution)
-		: origin(setOrigin), resolution(setResolution), direction(resolution*cos(verticalAngle)*cos(horizontalAngle), resolution*cos(verticalAngle)*sin(horizontalAngle),resolution*sin(verticalAngle)),
-		  castPosition(origin), castDistance(0)
-	{}
+		: origin(setOrigin)
+		, resolution(setResolution)
+		, direction(resolution * cos(verticalAngle) * cos(horizontalAngle),
+								resolution * cos(verticalAngle) * sin(horizontalAngle), resolution * sin(verticalAngle))
+		, castPosition(origin)
+		, castDistance(0)
+	{
+	}
 
-	void rayCast(const std::vector<Car>& cars, double minDistance, double maxDistance, pcl::PointCloud<pcl::PointXYZ>::Ptr& cloud, double slopeAngle, double sderr)
+	void rayCast(const std::vector<Car>& cars, double minDistance, double maxDistance,
+							 pcl::PointCloud<pcl::PointXYZ>::Ptr& cloud, double slopeAngle, double sderr)
 	{
 		// reset ray
 		castPosition = origin;
@@ -35,9 +43,8 @@ struct Ray
 
 		bool collision = false;
 
-		while(!collision && castDistance < maxDistance)
+		while (!collision && castDistance < maxDistance)
 		{
-
 			castPosition = castPosition + direction;
 			castDistance += resolution;
 
@@ -45,33 +52,31 @@ struct Ray
 			collision = (castPosition.z <= castPosition.x * tan(slopeAngle));
 
 			// check if there is any collisions with cars
-			if(!collision && castDistance < maxDistance)
+			if (!collision && castDistance < maxDistance)
 			{
-				for(Car car : cars)
+				for (Car car : cars)
 				{
 					collision |= car.checkCollision(castPosition);
-					if(collision)
+					if (collision)
 						break;
 				}
 			}
 		}
 
-		if((castDistance >= minDistance)&&(castDistance<=maxDistance))
+		if ((castDistance >= minDistance) && (castDistance <= maxDistance))
 		{
 			// add noise based on standard deviation error
-			double rx = ((double) rand() / (RAND_MAX));
-			double ry = ((double) rand() / (RAND_MAX));
-			double rz = ((double) rand() / (RAND_MAX));
-			cloud->points.push_back(pcl::PointXYZ(castPosition.x+rx*sderr, castPosition.y+ry*sderr, castPosition.z+rz*sderr));
+			double rx = ((double)rand() / (RAND_MAX));
+			double ry = ((double)rand() / (RAND_MAX));
+			double rz = ((double)rand() / (RAND_MAX));
+			cloud->points.push_back(
+					pcl::PointXYZ(castPosition.x + rx * sderr, castPosition.y + ry * sderr, castPosition.z + rz * sderr));
 		}
-			
 	}
-
 };
 
 struct Lidar
 {
-
 	std::vector<Ray> rays;
 	pcl::PointCloud<pcl::PointXYZ>::Ptr cloud;
 	std::vector<Car> cars;
@@ -83,7 +88,7 @@ struct Lidar
 	double sderr;
 
 	Lidar(std::vector<Car> setCars, double setGroundSlope)
-		: cloud(new pcl::PointCloud<pcl::PointXYZ>()), position(0,0,2.6)
+		: cloud(new pcl::PointCloud<pcl::PointXYZ>()), position(0, 0, 2.6)
 	{
 		// TODO:: set minDistance to 5 to remove points from roof of ego car
 		minDistance = 5;
@@ -97,18 +102,19 @@ struct Lidar
 		// TODO:: increase number of layers to 8 to get higher resoultion pcd
 		int numLayers = 8;
 		// the steepest vertical angle
-		double steepestAngle =  30.0*(-pi/180);
-		double angleRange = 26.0*(pi/180);
+		double steepestAngle = 30.0 * (-pi / 180);
+		double angleRange = 26.0 * (pi / 180);
 		// TODO:: set to pi/64 to get higher resoultion pcd
-		double horizontalAngleInc = pi/64;
+		double horizontalAngleInc = pi / 64;
 
-		double angleIncrement = angleRange/numLayers;
+		double angleIncrement = angleRange / numLayers;
 
-		for(double angleVertical = steepestAngle; angleVertical < steepestAngle+angleRange; angleVertical+=angleIncrement)
+		for (double angleVertical = steepestAngle; angleVertical < steepestAngle + angleRange;
+				 angleVertical += angleIncrement)
 		{
-			for(double angle = 0; angle <= 2*pi; angle+=horizontalAngleInc)
+			for (double angle = 0; angle <= 2 * pi; angle += horizontalAngleInc)
 			{
-				Ray ray(position,angle,angleVertical,resoultion);
+				Ray ray(position, angle, angleVertical, resoultion);
 				rays.push_back(ray);
 			}
 		}
@@ -116,23 +122,23 @@ struct Lidar
 
 	~Lidar()
 	{
-		// pcl uses boost smart pointers for cloud pointer so we don't have to worry about manually freeing the memory
+		// pcl uses boost smart pointers for cloud pointer so we don't have to worry
+		// about manually freeing the memory
 	}
 
 	pcl::PointCloud<pcl::PointXYZ>::Ptr scan()
 	{
 		cloud->points.clear();
 		auto startTime = std::chrono::steady_clock::now();
-		for(Ray ray : rays)
+		for (Ray ray : rays)
 			ray.rayCast(cars, minDistance, maxDistance, cloud, groundSlope, sderr);
 		auto endTime = std::chrono::steady_clock::now();
 		auto elapsedTime = std::chrono::duration_cast<std::chrono::milliseconds>(endTime - startTime);
-		cout << "ray casting took " << elapsedTime.count() << " milliseconds" << endl;
+		std::cout << "ray casting took " << elapsedTime.count() << " milliseconds" << std::endl;
 		cloud->width = cloud->points.size();
-		cloud->height = 1; // one dimensional unorganized point cloud dataset
+		cloud->height = 1;	// one dimensional unorganized point cloud dataset
 		return cloud;
 	}
-
 };
 
 #endif
